OpenGL 3D Game Tutorial 8: Model, View & Projection Matrices
Made a Camera which will allow me to move the whole world with WASD keys, as there were an actual camera.
This includes updating of the vertexShader to inlude uniform of the viewMatrix.
Also updated the rendere to test which triangle is closest and prio that rendering so that my cubes are actual cubes.