OpenGL 3D Game Tutorial 17: Multitexturing
vertexShader + fragmentShader has been updated.
MasterRenderer has been updated.

It is now possible to loadIn different textures instead of just 1.
OpenGL can save textures in different units for os, so we just have to tell java to save it in unit(x);
Make a texturePack that collects all the textures, and then have the GameLoop render it.